Fixed empty window bug when startup task was not provided
Enthought by default loads a saved layout including default task. If the
saved layout didn't exist, the startup task should have been loaded. But
I didn't provide a default startup task so you'd end up with a blank
window.

Note that the startup task idendifier is the 'id' trait of the
TaskFactory as defined in the Plugin.