Fix for OS X crash on exit: moved FontMapping* action to separate menu
Still don't know why, but including those actions in the same menu as the ANTIC font items crashed. I thought it was because I was using the same SMenu id which resulted in getting the actions added into the Antic Mode menu, but putting the Groups into the same SMenu still crashed. I thought it might be something to do with the Group ids being reused, so I changed the "a1", "a2", etc to something else, but that didn't work either. Ultimately, moving the FontMapping* stuff into its own SMenu with a different id worked.

Maybe because there were multiple generic Actions in different groups that got added into the same menu? Or because they were both radio buttons?